test: restore original integration assertions and snapshot
