+++ tests/integration.test.ts
@@ -1,312 +1,10 @@
-import { readdir, readFile, stat } from "node:fs/promises";
-import path from "node:path";
+import { get } from "svelte/store";
+import { describe, expect, test } from "vitest";
 
-import { noop } from "lodash/fp";
-import type { Moment } from "moment";
-import { type CachedMetadata, MetadataCache, type Vault } from "obsidian";
-import type { SListEntry, STask } from "obsidian-dataview";
-import { derived, get, writable } from "svelte/store";
-import { isNotVoid } from "typed-assert";
-import { describe, expect, test, vi } from "vitest";
-
-import { createUpdateHandler } from "../src/create-update-handler";
-import { dataviewChange } from "../src/redux/dataview/dataview-slice";
-import { initialState } from "../src/redux/global-slice";
-import { createReactor, type RootState } from "../src/redux/store";
-import type { DataviewFacade } from "../src/service/dataview-facade";
-import { TransactionWriter } from "../src/service/diff-writer";
-import { ListPropsParser } from "../src/service/list-props-parser";
-import type { PeriodicNotes } from "../src/service/periodic-notes";
-import { VaultFacade } from "../src/service/vault-facade";
-import type { WorkspaceFacade } from "../src/service/workspace-facade";
-import {
-  type DayPlannerSettings,
-  defaultSettingsForTests,
-} from "../src/settings";
-import { isLocal, type Task } from "../src/task-types";
+import { defaultSettingsForTests } from "../src/settings";
 import { EditMode } from "../src/ui/hooks/use-edit/types";
-import { useTasks } from "../src/ui/hooks/use-tasks";
-import { getOneLineSummary } from "../src/util/task-utils";
-
-import {
-  createInMemoryFile,
-  FakeDataviewFacade,
-  FakeMetadataCache,
-  FakePeriodicNotes,
-  FakeWorkspaceFacade,
-  getPathToDiff,
-  InMemoryVault,
-  type InMemoryFile,
-} from "./test-utils";
-
-const { join } = path.posix;
-
-const dailyNoteFileNames = ["2025-07-19", "2025-07-20", "2025-07-28"];
-
-const fixturesDirPath = "fixtures";
-const dumpPath = join(fixturesDirPath, "metadata-dump", "tasks.json");
-const fixtureVaultPath = join(fixturesDirPath, "fixture-vault");
-
-async function loadMetadataDump() {
-  const files = await readdir(fixtureVaultPath);
-
-  const inMemoryFiles = await Promise.all(
-    files.map(async (file) => {
-      const filePath = join(fixtureVaultPath, file);
-
-      const stats = await stat(filePath);
-
-      if (!stats.isFile()) {
-        throw new TypeError(
-          `Only files are supported in fixture vault, not this: ${filePath}`,
-        );
-      }
-
-      const contents = await readFile(filePath, "utf8");
-
-      return createInMemoryFile({ path: filePath, contents });
-    }),
-  );
-
-  const rawMetadataDump = await readFile(dumpPath, "utf-8");
-
-  const metadataDump = JSON.parse(rawMetadataDump);
-  const { tasks, lists, cachedMetadata } = metadataDump;
-
-  const pathToInMemoryFile = Object.fromEntries(
-    inMemoryFiles.map((it) => [it.path, it]),
-  );
-
-  const inMemoryDailyNotes = dailyNoteFileNames.map((it) => {
-    const path = join(fixtureVaultPath, `${it}.md`);
-    const file = pathToInMemoryFile[path];
-
-    isNotVoid(file, `There is no file for key: '${it}'`);
-
-    return { path, file, date: window.moment(it) };
-  });
-
-  return {
-    inMemoryFiles,
-    inMemoryDailyNotes,
-    tasks,
-    lists,
-    cachedMetadata,
-  };
-}
-
-function initTestServices(props: {
-  inMemoryFiles: InMemoryFile[];
-  inMemoryDailyNotes: { path: string; file: InMemoryFile; date: Moment }[];
-  tasks: STask[];
-  lists: SListEntry[];
-  cachedMetadata: Record<string, CachedMetadata>;
-}) {
-  const { inMemoryFiles, inMemoryDailyNotes, tasks, lists, cachedMetadata } =
-    props;
-
-  const periodicNotes = new FakePeriodicNotes(
-    inMemoryDailyNotes,
-  ) as unknown as PeriodicNotes;
-
-  const dataviewFacade = new FakeDataviewFacade({
-    tasks,
-    lists,
-  }) as unknown as DataviewFacade;
-
-  const metadataCache = new FakeMetadataCache(
-    cachedMetadata,
-  ) as unknown as MetadataCache;
-
-  const vault = new InMemoryVault(inMemoryFiles);
-
-  const getTasksApi = () => {
-    throw new Error("Can't access tasks API inside tests");
-  };
-
-  const vaultFacade = new VaultFacade(vault as unknown as Vault, getTasksApi);
-
-  const transactionWriter = new TransactionWriter(vaultFacade);
-
-  const workspaceFacade =
-    new FakeWorkspaceFacade() as unknown as WorkspaceFacade;
-
-  const listPropsParser = new ListPropsParser(
-    vault as unknown as Vault,
-    metadataCache,
-  );
-
-  return {
-    periodicNotes,
-    dataviewFacade,
-    metadataCache,
-    vault,
-    transactionWriter,
-    workspaceFacade,
-    vaultFacade,
-    listPropsParser,
-  };
-}
-
-async function setUp(props: {
-  visibleDays: string[];
-  settings?: DayPlannerSettings;
-}) {
-  const { visibleDays, settings = defaultSettingsForTests } = props;
-
-  const { inMemoryFiles, inMemoryDailyNotes, tasks, lists, cachedMetadata } =
-    await loadMetadataDump();
-
-  const {
-    periodicNotes,
-    dataviewFacade,
-    metadataCache,
-    vault,
-    transactionWriter,
-    workspaceFacade,
-    vaultFacade,
-    listPropsParser,
-  } = initTestServices({
-    inMemoryFiles,
-    inMemoryDailyNotes,
-    tasks,
-    lists,
-    cachedMetadata,
-  });
-
-  const visibleDaysStore = writable(visibleDays.map((it) => window.moment(it)));
-  const isOnline = writable(true);
-  const layoutReady = writable(true);
-  const settingsStore = writable(settings);
-  const currentTime = writable(window.moment());
-
-  const onEditCanceled = vi.fn();
-  const onEditConfirmed = vi.fn();
-  const defaultPreloadedStateForTests: Partial<RootState> = {
-    obsidian: {
-      ...initialState,
-      visibleDays,
-    },
-    settings: { settings },
-  };
-
-  const {
-    dispatch,
-    remoteTasks,
-    taskUpdateTrigger,
-    listProps,
-    pointerDateTime,
-  } = createReactor({
-    preloadedState: {
-      ...defaultPreloadedStateForTests,
-      obsidian: {
-        ...initialState,
-        visibleDays,
-      },
-    },
-    dataviewFacade,
-    listPropsParser,
-  });
-
-  inMemoryFiles.forEach(({ path }) => {
-    dispatch(dataviewChange(path));
-  });
-
-  const onUpdate = createUpdateHandler({
-    settings: () => settings,
-    transactionWriter,
-    vaultFacade,
-    periodicNotes,
-    onEditCanceled,
-    onEditConfirmed,
-    getTextInput: () => Promise.resolve("Text input"),
-    getConfirmationInput: () => Promise.resolve(true),
-  });
-
-  const {
-    tasksWithActiveClockProps,
-    getDisplayedTasksWithClocksForTimeline,
-    tasksWithTimeForToday,
-    editContext,
-    newlyStartedTasks,
-  } = useTasks({
-    onUpdate,
-    onEditAborted: () => {},
-    periodicNotes,
-    dataviewFacade,
-    metadataCache,
-    workspaceFacade,
-    isOnline,
-    visibleDays: visibleDaysStore,
-    layoutReady,
-    debouncedTaskUpdateTrigger: taskUpdateTrigger,
-    dataviewChange: taskUpdateTrigger,
-    settingsStore,
-    currentTime,
-    pointerDateTime,
-    remoteTasks,
-    listProps,
-  });
-
-  const allTasks = derived(
-    editContext.dayToDisplayedTasks,
-    ($dayToDisplayedTasks) => {
-      return Object.values($dayToDisplayedTasks).flatMap(
-        ({ withTime, noTime }) => withTime.concat(noTime),
-      );
-    },
-  );
-
-  // this prevents the store from resetting;
-  allTasks.subscribe(noop);
-
-  function moveCursorTo(
-    dateTime: Moment,
-    type: "date" | "dateTime" = "dateTime",
-  ) {
-    pointerDateTime.set({
-      dateTime,
-      type,
-    });
-  }
-
-  function findTask(predicate: (task: Task) => boolean) {
-    const found = get(allTasks).filter(isLocal).find(predicate);
-
-    isNotVoid(found, `Task not found`);
-
-    return found;
-  }
-
-  function findByText(text: string) {
-    return findTask((it) => getOneLineSummary(it).includes(text));
-  }
-
-  await vi.waitFor(() => {
-    const areTasksLoaded = get(allTasks).length > 0;
-    const areClocksLoaded = get(tasksWithActiveClockProps).length > 0;
-
-    expect(areTasksLoaded || areClocksLoaded).toBeTruthy();
-  });
-
-  return {
-    dispatch,
-    tasksWithActiveClockProps,
-    getDisplayedTasksWithClocksForTimeline,
-    tasksWithTimeForToday,
-    editContext,
-    newlyStartedTasks,
-    periodicNotes,
-    moveCursorTo,
-    onUpdate,
-    vault,
-    findTask,
-    findByText,
-    allTasks,
-    transactionWriter,
-    currentTime,
-  };
-}
+import { getPathToDiff } from "./test-utils";
+import { setUp } from "./integration/init-test-system";
 
 describe("Clocks", () => {
   test("Reads log data", async () => {
